functions_project/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
azure-monitor-opentelemetry

azure-functions

dotenv

tiktoken

pillow

azure-cognitiveservices-speech